+++ onnx/backend/base.py
@@ -21,6 +21,7 @@ class Device(object):
     syntax: device_type:device_id(optional)
     example: 'CPU', 'CUDA', 'CUDA:1'
     '''
+
     def __init__(self, device):
         options = device.split(':')
         self.type = getattr(DeviceType, options[0])
@@ -34,6 +35,7 @@ def namedtupledict(typename, field_names, *args, **kwargs):
     # Some output names are invalid python identifier, e.g. "0"
     kwargs.setdefault('rename', True)
     data = namedtuple(typename, field_names, *args, **kwargs)
+
     def getitem(self, key):
         if isinstance(key, six.string_types):
             key = field_names_map[key]
@@ -57,7 +59,14 @@ class Backend(object):
         return cls.prepare(model, device, **kwargs).run(inputs)
 
     @classmethod
-    def run_node(cls, node, inputs, device='CPU'):
+    def run_node(cls, node, inputs, device='CPU', outputs_info=None, **kwargs):
+        '''Simple run one operator and return the results.
+        Args:
+            outputs_info: a list of tuples, which contains the element type and
+            shape of each output. First element of the tuple is the dtype, and
+            the second element is the shape. More use case can be found in
+            https://github.com/onnx/onnx/blob/master/onnx/backend/test/runner/__init__.py
+        '''
         onnx.checker.check_node(node)
 
     @classmethod