+++ onnx/defs/tensor/defs.cc
@@ -15,10 +15,10 @@ in the 'DataType' enum field in the TensorProto message.
 NOTE: Casting to and from strings is not supported yet.
 )DOC")
     .Attr(
-          "to",
-          "The data type to which the elements of the input tensor are cast."
-          "Strictly must be one of the types from DataType enum in TensorProto",
-          AttributeProto::STRING)
+        "to",
+        "The data type to which the elements of the input tensor are cast."
+        "Strictly must be one of the types from DataType enum in TensorProto",
+        AttributeProto::STRING)
     .Input(0, "input", "Input tensor to be cast.", "T1")
     .Output(
         0,
@@ -26,28 +26,56 @@ NOTE: Casting to and from strings is not supported yet.
         "Output tensor with the same shape as input with type "
         "specified by the 'to' argument",
         "T2")
-    .TypeConstraint("T1", { "tensor(float16)", "tensor(float)", "tensor(double)" },
-            "Constrain input types to float tensors.")
-    .TypeConstraint("T2", { "tensor(float16)", "tensor(float)", "tensor(double)" },
-            "Constrain output types to float tensors.");
-
+    .TypeConstraint(
+        "T1",
+        {"tensor(float16)",
+         "tensor(float)",
+         "tensor(double)",
+         "tensor(int8)",
+         "tensor(int16)",
+         "tensor(int32)",
+         "tensor(int64)",
+         "tensor(uint8)",
+         "tensor(uint16)",
+         "tensor(uint32)",
+         "tensor(uint64)",
+         "tensor(bool)"},
+        "Constrain input types. Casting from strings and complex are not supported.")
+    .TypeConstraint(
+        "T2",
+        {"tensor(float16)",
+         "tensor(float)",
+         "tensor(double)",
+         "tensor(int8)",
+         "tensor(int16)",
+         "tensor(int32)",
+         "tensor(int64)",
+         "tensor(uint8)",
+         "tensor(uint16)",
+         "tensor(uint32)",
+         "tensor(uint64)",
+         "tensor(bool)"},
+        "Constrain output types. Casting to strings and complex are not supported.");
 
 ONNX_OPERATOR_SCHEMA(Reshape)
     .AllowConsumed({{0, 0}})
     .SetDoc(R"DOC(
 Reshape the input tensor similar to numpy.reshape.
 
-It takes a tensor as input and an argument `shape`. It outputs the reshaped tensor.
+First input is the data tensor, second input is a shape tensor which specifies the output shape. It outputs the reshaped tensor.
 
 At most one dimension of the new shape can be -1. In this case, the value is
 inferred from the size of the tensor and the remaining dimensions. A dimension
 could also be 0, in which case the actual dimension value is unchanged (i.e. taken
 from the input tensor).)DOC")
-    .Attr("shape", "New shape", AttributeProto::INTS, OPTIONAL)
     .Input(0, "data", "An input tensor.", "T")
+    .Input(1, "shape", "Specified shape for output.", "tensor(int64)")
+    .SinceVersion(5)
     .Output(0, "reshaped", "Reshaped data.", "T")
-    .TypeConstraint("T", { "tensor(float16)", "tensor(float)", "tensor(double)" },
-	            "Constrain input and output types to float tensors.");
+    .TypeConstraint(
+        "T",
+        {"tensor(float16)", "tensor(float)", "tensor(double)"},
+        "Constrain input and output types to float tensors.");
 
 ONNX_OPERATOR_SCHEMA(Shape)
     .SetDoc(R"DOC(
@@ -55,14 +83,23 @@ Takes a tensor as input and outputs an 1D int64 tensor containing the shape of t
 )DOC")
     .Input(0, "data", "An input tensor.", "T")
     .Output(0, "shape", "Shape of the input tensor", "T1")
-    .TypeConstraint("T", { "tensor(float16)",  "tensor(float)", "tensor(double)",
-                           "tensor(int8)", "tensor(int16)",  "tensor(int32)", "tensor(int64)",
-                           "tensor(uint8)", "tensor(uint16)",
-                           "tensor(bool)" },
-                "Input tensor can be of arbitrary type.")
-    .TypeConstraint("T1", { "tensor(int64)"},
-	            "Constrains output to int64 tensor.");
-
+    .TypeConstraint(
+        "T",
+        {"tensor(float16)",
+         "tensor(float)",
+         "tensor(double)",
+         "tensor(int8)",
+         "tensor(int16)",
+         "tensor(int32)",
+         "tensor(int64)",
+         "tensor(uint8)",
+         "tensor(uint16)",
+         "tensor(bool)"},
+        "Input tensor can be of arbitrary type.")
+    .TypeConstraint(
+        "T1",
+        {"tensor(int64)"},
+        "Constrains output to int64 tensor.");
 
 ONNX_OPERATOR_SCHEMA(Size)
     .SetDoc(R"DOC(
@@ -70,40 +107,52 @@ Takes a tensor as input and outputs a int64 scalar that equals to the total numb
 )DOC")
     .Input(0, "data", "An input tensor.", "T")
     .Output(0, "size", "Total number of elements of the input tensor", "T1")
-    .TypeConstraint("T", { "tensor(float16)",  "tensor(float)", "tensor(double)",
-                           "tensor(int8)", "tensor(int16)",  "tensor(int32)", "tensor(int64)",
-                           "tensor(uint8)", "tensor(uint16)",
-                           "tensor(bool)" },
-                "Input tensor can be of arbitrary type.")
-    .TypeConstraint("T1", { "int64" },
-	            "Constrains output to int64 scalar.");
-
+    .TypeConstraint(
+        "T",
+        {"tensor(float16)",
+         "tensor(float)",
+         "tensor(double)",
+         "tensor(int8)",
+         "tensor(int16)",
+         "tensor(int32)",
+         "tensor(int64)",
+         "tensor(uint8)",
+         "tensor(uint16)",
+         "tensor(bool)"},
+        "Input tensor can be of arbitrary type.")
+    .TypeConstraint("T1", {"int64"}, "Constrains output to int64 scalar.");
 
 ONNX_OPERATOR_SCHEMA(Concat)
-.SinceVersion(4)
-.Attr("axis",
-    "Which axis to concat on",
-    AttributeProto::INT)
+    .SinceVersion(4)
+    .Attr("axis", "Which axis to concat on", AttributeProto::INT)
     .SetDoc("Concatenate a list of tensors into a single tensor")
-    .Input(0, "inputs", "List of tensors for concatenation", "T", OpSchema::Variadic)
+    .Input(
+        0,
+        "inputs",
+        "List of tensors for concatenation",
+        "T",
+        OpSchema::Variadic)
     .Output(0, "concat_result", "Concatenated tensor", "T")
-    .TypeConstraint("T", { "tensor(float16)", "tensor(float)", "tensor(double)" },
+    .TypeConstraint(
+        "T",
+        {"tensor(float16)", "tensor(float)", "tensor(double)"},
         "Constrain output types to float tensors.");
 
 ONNX_OPERATOR_SCHEMA(Split)
     .SinceVersion(2)
     .Input(0, "input", "The tensor to split", "T")
-    .Output(0, "outputs", "One or more outputs forming list of tensors after splitting", "T", OpSchema::Variadic)
-    .TypeConstraint("T", { "tensor(float16)", "tensor(float)", "tensor(double)" },
-            "Constrain input types to float tensors.")
-    .Attr("axis",
-          "Which axis to split on",
-          AttributeProto::INT,
-          OPTIONAL)
-    .Attr("split",
-          "length of each output",
-          AttributeProto::INTS,
-          OPTIONAL)
+    .Output(
+        0,
+        "outputs",
+        "One or more outputs forming list of tensors after splitting",
+        "T",
+        OpSchema::Variadic)
+    .TypeConstraint(
+        "T",
+        {"tensor(float16)", "tensor(float)", "tensor(double)"},
+        "Constrain input types to float tensors.")
+    .Attr("axis", "Which axis to split on", AttributeProto::INT, OPTIONAL)
+    .Attr("split", "length of each output", AttributeProto::INTS, OPTIONAL)
     .SetDoc(R"DOC(Split a tensor into a list of tensors, along the specified
 'axis'. Lengths of the parts can be specified using argument 'split'.
 Otherwise, the tensor is split to equal sized parts.
@@ -153,21 +202,26 @@ Example 2:
 
 )DOC")
     .Input(0, "data", "Tensor of data to extract slices from.", "T")
-    .Attr("axes",
-          "Axes that `starts` and `ends` apply to. "
-          "It's optional. If not present, will be treated as "
-          "[0, 1, ..., len(`starts`) - 1].",
-          AttributeProto::INTS,
-          OPTIONAL)
-    .Attr("starts",
-          "Starting indices of corresponding axis in `axes`",
-          AttributeProto::INTS)
-    .Attr("ends",
-          "Ending indices (exclusive) of corresponding axis in axes`",
-          AttributeProto::INTS)
+    .Attr(
+        "axes",
+        "Axes that `starts` and `ends` apply to. "
+        "It's optional. If not present, will be treated as "
+        "[0, 1, ..., len(`starts`) - 1].",
+        AttributeProto::INTS,
+        OPTIONAL)
+    .Attr(
+        "starts",
+        "Starting indices of corresponding axis in `axes`",
+        AttributeProto::INTS)
+    .Attr(
+        "ends",
+        "Ending indices (exclusive) of corresponding axis in axes`",
+        AttributeProto::INTS)
     .Output(0, "output", "Sliced data tensor.", "T")
-    .TypeConstraint("T", { "tensor(float16)", "tensor(float)", "tensor(double)" },
-            "Constrain input and output types to float tensors.");
+    .TypeConstraint(
+        "T",
+        {"tensor(float16)", "tensor(float)", "tensor(double)"},
+        "Constrain input and output types to float tensors.");
 
 ONNX_OPERATOR_SCHEMA(Transpose)
     .SetDoc(R"DOC(
@@ -175,15 +229,18 @@ Transpose the input tensor similar to numpy.transpose. For example, when
 perm=(1, 0, 2), given an input tensor of shape (1, 2, 3), the output shape
 will be (2, 1, 3).
 )DOC")
-    .Attr("perm",
-          "A list of integers. By default, reverse the dimensions, "
-          "otherwise permute the axes according to the values given.",
-          AttributeProto::INTS,
-          OPTIONAL)
+    .Attr(
+        "perm",
+        "A list of integers. By default, reverse the dimensions, "
+        "otherwise permute the axes according to the values given.",
+        AttributeProto::INTS,
+        OPTIONAL)
     .Input(0, "data", "An input tensor.", "T")
     .Output(0, "transposed", "Transposed output.", "T")
-    .TypeConstraint("T", { "tensor(float16)", "tensor(float)", "tensor(double)" },
-            "Constrain input and output types to float tensors.");
+    .TypeConstraint(
+        "T",
+        {"tensor(float16)", "tensor(float)", "tensor(double)"},
+        "Constrain input and output types to float tensors.");
 
 ONNX_OPERATOR_SCHEMA(Gather)
     .SetDoc(R"DOC(
@@ -253,23 +310,26 @@ Example 2:
         "Constrain indices to integer types");
 
 ONNX_OPERATOR_SCHEMA(Squeeze)
-    .Attr("axes",
-          "List of positive integers, indicate the dimensions to squeeze.",
-          AttributeProto::INTS)
+    .Attr(
+        "axes",
+        "List of positive integers, indicate the dimensions to squeeze.",
+        AttributeProto::INTS)
     .SetDoc(R"DOC(
 Remove single-dimensional entries from the shape of a tensor.
 Takes a  parameter `axes` with a list of axes to squeeze.
 )DOC")
     .Input(0, "data", "Tensors with at least max(dims) dimensions.", "T")
     .Output(0, "squeezed", "Reshaped tensor with same data as input.", "T")
-    .TypeConstraint("T", { "tensor(float16)", "tensor(float)", "tensor(double)" },
-            "Constrain input and output types to float tensors.");
+    .TypeConstraint(
+        "T",
+        {"tensor(float16)", "tensor(float)", "tensor(double)"},
+        "Constrain input and output types to float tensors.");
 
 ONNX_OPERATOR_SCHEMA(Unsqueeze)
     .Attr(
-      "axes",
-      "List of positive integers, indicate the dimensions to be inserted",
-      AttributeProto::INTS)
+        "axes",
+        "List of positive integers, indicate the dimensions to be inserted",
+        AttributeProto::INTS)
     .SetDoc(R"DOC(
 Insert single-dimensional entries to the shape of a tensor.
 Takes one required argument `axes`, a list of dimensions that will be inserted.
@@ -281,27 +341,32 @@ Dimension indices in `axes` are as seen in the output tensor. For example:
 )DOC")
     .Input(0, "data", "Original tensor", "T")
     .Output(0, "expanded", "Reshaped tensor with same data as input.", "T")
-    .TypeConstraint("T", { "tensor(float16)", "tensor(float)", "tensor(double)" },
-            "Constrain input and output types to float tensors.");
+    .TypeConstraint(
+        "T",
+        {"tensor(float16)", "tensor(float)", "tensor(double)"},
+        "Constrain input and output types to float tensors.");
 
 ONNX_OPERATOR_SCHEMA(Pad)
     .SinceVersion(2)
-    .Attr("pads",
-          "List of integers indicate the padding element count at the "
-          "beginning and end of each axis, for 2D it is the number of pixel. "
-          "`pads` rank should be double of the input's rank. `pads` format should be as follow "
-          "[x1_begin, x2_begin...x1_end, x2_end,...], where xi_begin the number of pixels "
-          "added at the beginning of axis `i` and xi_end, the number of pixels added at "
-          "the end of axis `i`.",
-          AttributeProto::INTS)
-    .Attr("mode",
-          "Three modes: constant(default), reflect, edge",
-          AttributeProto::STRING,
-          std::string("constant"))
-    .Attr("value",
-          "One float, indicates the value to be filled, default is 0",
-          AttributeProto::FLOAT,
-          0.0f)
+    .Attr(
+        "pads",
+        "List of integers indicate the padding element count at the "
+        "beginning and end of each axis, for 2D it is the number of pixel. "
+        "`pads` rank should be double of the input's rank. `pads` format should be as follow "
+        "[x1_begin, x2_begin...x1_end, x2_end,...], where xi_begin the number of pixels "
+        "added at the beginning of axis `i` and xi_end, the number of pixels added at "
+        "the end of axis `i`.",
+        AttributeProto::INTS)
+    .Attr(
+        "mode",
+        "Three modes: constant(default), reflect, edge",
+        AttributeProto::STRING,
+        std::string("constant"))
+    .Attr(
+        "value",
+        "One float, indicates the value to be filled, default is 0",
+        AttributeProto::FLOAT,
+        0.0f)
     .SetDoc(R"DOC(
 Given `data` tensor, pads, mode, and value.
 
@@ -325,61 +390,77 @@ Example:
 )DOC")
     .Input(0, "data", "Input tensor.", "T")
     .Output(0, "output", "Tensor after padding.", "T")
-    .TypeConstraint("T", { "tensor(float16)", "tensor(float)", "tensor(double)" },
-            "Constrain input and output types to float tensors.");
+    .TypeConstraint(
+        "T",
+        {"tensor(float16)", "tensor(float)", "tensor(double)"},
+        "Constrain input and output types to float tensors.");
 
 ONNX_OPERATOR_SCHEMA(SpaceToDepth)
-    .Attr("blocksize",
-          "Blocks of [blocksize, blocksize] are moved.",
-          AttributeProto::INT,
-          OPTIONAL)
-    .SetDoc(R"DOC(SpaceToDepth rearranges blocks of spatial data into depth. More specifically,
+    .Attr(
+        "blocksize",
+        "Blocks of [blocksize, blocksize] are moved.",
+        AttributeProto::INT,
+        OPTIONAL)
+    .SetDoc(
+        R"DOC(SpaceToDepth rearranges blocks of spatial data into depth. More specifically,
 this op outputs a copy of the input tensor where values from the height and width dimensions
 are moved to the depth dimension.
 )DOC")
-    .Input(0,
-           "input",
-           "Input tensor of [N,C,H,W], where N is the batch axis, C is the channel or depth"
-           ", H is the height and W is the width.", "T")
-    .Output(0,
-            "output",
-            "Output tensor of [N, C * blocksize * blocksize, H/blocksize, W/blocksize].", "T")
-    .TypeConstraint("T", { "tensor(float16)", "tensor(float)", "tensor(double)" },
-            "Constrain input types to float tensors.");
+    .Input(
+        0,
+        "input",
+        "Input tensor of [N,C,H,W], where N is the batch axis, C is the channel or depth"
+        ", H is the height and W is the width.",
+        "T")
+    .Output(
+        0,
+        "output",
+        "Output tensor of [N, C * blocksize * blocksize, H/blocksize, W/blocksize].",
+        "T")
+    .TypeConstraint(
+        "T",
+        {"tensor(float16)", "tensor(float)", "tensor(double)"},
+        "Constrain input types to float tensors.");
 
 ONNX_OPERATOR_SCHEMA(DepthToSpace)
-    .Attr("blocksize",
-          "Blocks of [blocksize, blocksize] are moved.",
-          AttributeProto::INT,
-          OPTIONAL)
-    .SetDoc(R"DOC(DepthToSpace rearranges (permutes) data from depth into blocks of spatial data.
+    .Attr(
+        "blocksize",
+        "Blocks of [blocksize, blocksize] are moved.",
+        AttributeProto::INT,
+        OPTIONAL)
+    .SetDoc(
+        R"DOC(DepthToSpace rearranges (permutes) data from depth into blocks of spatial data.
 This is the reverse transformation of SpaceToDepth. More specifically, this op outputs a copy of
 the input tensor where values from the depth dimension are moved in spatial blocks to the height
 and width dimensions.
 )DOC")
-    .Input(0,
-           "input",
-           "Input tensor of [N,C,H,W], where N is the batch axis, C is the channel or depth"
-           ", H is the height and W is the width.", "T")
-    .Output(0,
-            "output",
-            "Output tensor of [N, C/(blocksize * blocksize), H * blocksize, W * blocksize].", "T")
-    .TypeConstraint("T", { "tensor(float16)", "tensor(float)", "tensor(double)" },
-            "Constrain input types to float tensors.");
+    .Input(
+        0,
+        "input",
+        "Input tensor of [N,C,H,W], where N is the batch axis, C is the channel or depth"
+        ", H is the height and W is the width.",
+        "T")
+    .Output(
+        0,
+        "output",
+        "Output tensor of [N, C/(blocksize * blocksize), H * blocksize, W * blocksize].",
+        "T")
+    .TypeConstraint(
+        "T",
+        {"tensor(float16)", "tensor(float)", "tensor(double)"},
+        "Constrain input types to float tensors.");
 
 ONNX_OPERATOR_SCHEMA(Tile)
     .SetDoc(R"DOC(Repeat the elements of a tensor along an axis.)DOC")
-    .Input(0,
-           "input",
-           "Input tensor of any shape.", "T")
-    .Input(1,
-           "tiles",
-           "Number of repeated copies to make of the input tensor.", "T")
-    .Input(2,
-           "axis",
-           "Axis along which to repeat.", "T")
-    .Output(0,
-            "output",
-            "Output tensor of same shape and type as input.", "T")
-    .TypeConstraint("T", { "tensor(float16)", "tensor(float)", "tensor(double)" },
-            "Constrain input types to float tensors.");
+    .Input(0, "input", "Input tensor of any shape.", "T")
+    .Input(
+        1,
+        "tiles",
+        "Number of repeated copies to make of the input tensor.",
+        "T")
+    .Input(2, "axis", "Axis along which to repeat.", "T")
+    .Output(0, "output", "Output tensor of same shape and type as input.", "T")
+    .TypeConstraint(
+        "T",
+        {"tensor(float16)", "tensor(float)", "tensor(double)"},
+        "Constrain input types to float tensors.");